feat: reorganize component structure
