+++ package.json
@@ -18,6 +18,7 @@
       "@reduxjs/toolkit": "^2.5.0",
       "axios": "^1.7.9",
       "dompurify": "^3.2.4",
+      "gsap": "^3.12.7",
       "lenis": "^1.1.19",
       "react": "^18.3.1",
       "react-dom": "^18.3.1",
@@ -30,7 +31,7 @@
       "react-slick": "^0.30.3",
       "react-toastify": "^11.0.3",
       "slick-carousel": "^1.8.1",
-      "sweetalert2": "^11.15.10",
+      "sweetalert2": "^11.6.13",
       "swiper": "^11.2.1",
       "yup": "^1.6.1"
    },
