+++ src/components/Templates/about/Teachers.jsx
@@ -1,7 +1,7 @@
 import React, { Suspense, useState, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
-import useTeacher from '../../hooks/useTeacher';
+import useTeacher from '../../../hooks/useTeacher';
 // import {Swiper} from 'swiper';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import 'swiper/css';
