+++ src/components/Templates/album/Album_Card.jsx
