+++ src/components/Templates/gallery/GalleryCard.jsx
