+++ src/components/Templates/home/Delete_Album.js
