+++ src/components/Templates/home/Testimonials.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
-import useTestimonial from '../../hooks/useTestimoniyal';
+import useTestimonial from '../../../hooks/useTestimoniyal';
 import { useSelector } from 'react-redux';
 
 const Testimonials = () => {
