+++ src/pages/About.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import { Helmet } from 'react-helmet-async';
-import History from '../components/aboutComponents/History';
-import AboutHero from '../components/aboutComponents/AboutHero';
-import Teachers from '../components/aboutComponents/Teachers';
-import Courses from '../components/aboutComponents/Courses';
-import CampusFacilities from '../components/aboutComponents/CampusFacilities';
+import History from '../components/Templates/about/History';
+import AboutHero from '../components/Templates/about/AboutHero';
+import Teachers from '../components/Templates/about/Teachers';
+import Courses from '../components/Templates/about/Courses';
+import CampusFacilities from '../components/Templates/about/CampusFacilities';
 import { useSelector } from 'react-redux';
 import useAbout from '../hooks/useAbout';
-import Loader from '../components/Loader';
+import Loader from '../components/Molecules/Loader/Loader';
+
 
 const About = () => {
    useAbout();
